# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  domain.SpecProvenance:
    properties:
      author:
        type: string
      createdAt:
        type: string
      id:
        type: string
      project:
        type: string
    type: object
  handler.artifactRefDTO:
    properties:
      capability:
        type: string
      changeName:
        type: string
      kind:
        type: string
    type: object
...
  handler.createSpecRequest:
    properties:
      capability:
        type: string
      projectId:
        type: integer
      purpose:
        type: string
      requirements:
        items:
          $ref: '#/definitions/handler.specRequirementDTO'
        type: array
    required:
    - capability
    - projectId
    type: object
  handler.debtResponse:
    properties:
      body:
        type: string
      createdAt:
        type: string
      id:
        type: string
      link:
        $ref: '#/definitions/handler.artifactRefDTO'
      projectId:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  handler.dependencyRequest:
    properties:
      targetCapability:
        type: string
      targetProjectId:
        type: integer
    required:
    - targetCapability
    - targetProjectId
    type: object
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.generateSpecRequest:
    properties:
      description:
        type: string
      providerId:
        description: 0 = use the configured default provider
        type: integer
    required:
    - description
    type: object
...
  handler.reviewResponse:
    properties:
      body:
        type: string
      createdAt:
        type: string
      id:
        type: string
      link:
        $ref: '#/definitions/handler.artifactRefDTO'
      projectId:
        type: integer
      title:
        type: string
    type: object
  handler.specDependencyDTO:
    properties:
      available:
        type: boolean
      capability:
        type: string
      projectName:
        type: string
      projectPath:
        type: string
    type: object
  handler.specRequirementDTO:
    properties:
      body:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  handler.specResponse:
    properties:
      capability:
        type: string
      dependedBy:
        items:
          $ref: '#/definitions/handler.specDependencyDTO'
        type: array
      dependsOn:
        items:
          $ref: '#/definitions/handler.specDependencyDTO'
        type: array
      id:
        type: string
      linkedDebts:
        items:
          $ref: '#/definitions/handler.debtResponse'
        type: array
      linkedReviews:
        items:
          $ref: '#/definitions/handler.reviewResponse'
        type: array
      projectId:
        type: integer
      projectName:
        type: string
      provenance:
        $ref: '#/definitions/domain.SpecProvenance'
      purpose:
        type: string
      requirements:
        items:
          $ref: '#/definitions/handler.specRequirementDTO'
        type: array
    type: object
...
  handler.updateSpecRequest:
    properties:
      purpose:
        type: string
      requirements:
        items:
          $ref: '#/definitions/handler.specRequirementDTO'
        type: array
    type: object
info:
  contact: {}
  description: REST + SSE API for the OpenSpec Studio backend.
  title: OpenSpec Studio API
  version: "1.0"
paths:
...
  /specs:
    get:
      description: Lists every spec across every registered, available project (or
        a single project via ?projectId=), with resolved dependency links.
      parameters:
      - description: Filter to a single registered project
        in: query
        name: projectId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.specResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: List specs
      tags:
      - specs
    post:
      consumes:
      - application/json
      parameters:
      - description: New spec
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createSpecRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.specResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: spec already exists
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create a spec
      tags:
      - specs
  /specs/{id}:
    delete:
      description: Deletes a spec. If other specs depend on it and force=true is not
        set, returns 409 with the list of dependent specs instead of deleting.
      parameters:
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      - description: Delete despite dependents
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: dependent specs
          schema:
            items:
              $ref: '#/definitions/handler.specResponse'
            type: array
      summary: Delete a spec
      tags:
      - specs
    get:
      parameters:
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.specResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get a spec
      tags:
      - specs
    patch:
      consumes:
      - application/json
      parameters:
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      - description: Updated purpose/requirements
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateSpecRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.specResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Update a spec
      tags:
      - specs
  /specs/{id}/dependencies:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      - description: Dependency target
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.dependencyRequest'
      responses:
        "204":
          description: No Content
      summary: Remove a spec dependency
      tags:
      - specs
    post:
      consumes:
      - application/json
      parameters:
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      - description: Dependency target
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.dependencyRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: self-dependency rejected
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Add a spec dependency
      tags:
      - specs
  /specs/generate:
    post:
      consumes:
      - application/json
      description: Drafts a spec from a short description without writing anything
        to disk. Uses the default AI provider, or an explicit providerId override,
        falling back to a deterministic template if none is configured or the provider
        call fails.
      parameters:
      - description: Description (and optional provider override)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.generateSpecRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.specResponse'
      summary: Generate a draft spec
      tags:
      - specs
swagger: "2.0"